Update blueprint.py
removed unnecessary int()

+++ broadbean/blueprint.py
@@ -762,7 +762,7 @@ def _subelementBuilder(blueprint: BluePrint, SR: int,
     intdurations = np.zeros(len(newdurations), dtype=int)
 
     for ii, dur in enumerate(newdurations):
-        int_dur = int(round(dur*SR))
+        int_dur = round(dur*SR)
         if int_dur < 2:
             raise SegmentDurationError('Too short segment detected! '
                                        'Segment "{}" at position {} '
